Terms and conditions update
I've added a new paragraph to the first section to make it clear that a user setting up an account is agreeing to accept documents electronically from the court. I haven't been able to preview this so it'd be good if someone could show me it and then it can be checked as normal.