Ronan civil procedure rules content (#589)
* Content changes based on Civil Procedure rules feedback

1. I've changed the two eligibility questions which mention an address so they say "postal address" instead - this is apparently important.
2. I've added "this is the statutory rate" to the help text when we talk about 8% interest - on two different pages. This is apparently important and insisted on despite the lack of a user need.

* Terms and conditions update

I've added a new paragraph to the first section to make it clear that a user setting up an account is agreeing to accept documents electronically from the court. I haven't been able to preview this so it'd be good if someone could show me it and then it can be checked as normal.

* Mediation content link

I've added a link explaining how mediation works so a claimant can decide whether to accept the defendant's offer of mediation.

* Privacy policy content

I've made a number of changes to this based on instruction from the head of content, Fran. I've hopefully not introduced any errors, tried to follow the code as best I could. Thanks. It'd be nice to see how this looks before we publish it.

* Fix external link and tidy up code

* Update claimRejected.njk

* Capitilise MOJ