Minor content change
Based on research I've changed the hint text in the first question to make it match the other two hint texts - it explains the question instead of telling the user about a future step.

I've added 'full' to the second question's hint text to match my hint for the first.

I've changed 'you have' to 'you've' in the third question to meet GOV.UK style.