Make minor content changes (#615)
1. I changed the dash, adding 'you said' to this line so it says 'telling them when and how you said you paid the claim' - the previous wording implies the service agrees their defence is true.
2. I changed 'your offer has been sent' to 'we've sent your offer' - just because our style is make it clear who has done something and not use the passive voice.

Changed the other instance of 'your offer has been sent' to 'we've sent your offer'

I've updated this because it's not factually correct to say that the agreement replaces the claim. I've also used a more active wording 'we've asked...' instead of 'will be asked'

Based on research I've changed the hint text in the first question to make it match the other two hint texts - it explains the question instead of telling the user about a future step.

I've added 'full' to the second question's hint text to match my hint for the first.

I've changed 'you have' to 'you've' in the third question to meet GOV.UK style.

User research shows users don't always think of their claim as monetary in the early stages of making it. I've edited this page to make it less specifically about money and more about a dispute that needs to be resolved.