Removing form names
We've decided to, in general, remove the form names as they don't mean anything to users.

This meant me going through each ineligibility screen and changing 'n1 form' to 'paper form'. In the process I've had to make other content changes to make it make sense. I've also made our links all contain verbs, as far as possible, as that's a general GDS standard and is more accessible.

Hoping I've not introduced any errors but this was fairly detailed changes so apologies if so.